dsearch -> search.faiss (#1028)
moved dsearch into search.faiss; renamed SimpleDenseSearcher into FaissSearcher; added fake API for original access
