+++ integrations/dense/test_tct_colbert.py
@@ -21,7 +21,7 @@ import socket
 import unittest
 
 from integrations.utils import clean_files, run_command, parse_score
-from pyserini.dsearch import QueryEncoder
+from pyserini.search import QueryEncoder
 from pyserini.search import get_topics
 
 
@@ -39,7 +39,7 @@ class TestSearchIntegration(unittest.TestCase):
     def test_msmarco_passage_tct_colbert_bf_otf(self):
         output_file = 'test_run.msmarco-passage.tct_colbert.bf-otf.tsv'
         self.temp_files.append(output_file)
-        cmd1 = f'python -m pyserini.dsearch --topics msmarco-passage-dev-subset \
+        cmd1 = f'python -m pyserini.search.faiss --topics msmarco-passage-dev-subset \
                              --index msmarco-passage-tct_colbert-bf \
                              --encoder castorini/tct_colbert-msmarco \
                              --batch-size {self.batch_size} \
@@ -57,7 +57,7 @@ class TestSearchIntegration(unittest.TestCase):
     def test_msmarco_passage_tct_colbert_hnsw_otf(self):
         output_file = 'test_run.msmarco-passage.tct_colbert.hnsw-otf.tsv'
         self.temp_files.append(output_file)
-        cmd1 = f'python -m pyserini.dsearch --topics msmarco-passage-dev-subset \
+        cmd1 = f'python -m pyserini.search.faiss --topics msmarco-passage-dev-subset \
                              --index msmarco-passage-tct_colbert-hnsw \
                              --encoder castorini/tct_colbert-msmarco \
                              --output {output_file} \
@@ -114,7 +114,7 @@ class TestSearchIntegration(unittest.TestCase):
     def test_msmarco_doc_tct_colbert_bf_otf(self):
         output_file = 'test_run.msmarco-doc.passage.tct_colbert-otf.txt'
         self.temp_files.append(output_file)
-        cmd1 = f'python -m pyserini.dsearch --topics msmarco-doc-dev \
+        cmd1 = f'python -m pyserini.search.faiss --topics msmarco-doc-dev \
                              --index msmarco-doc-tct_colbert-bf \
                              --encoder castorini/tct_colbert-msmarco \
                              --output {output_file} \
